backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
PyMySQL>=1.1
cryptography>=42.0
razorpay>=1.4
python-dotenv>=1.0
